SitioClinicas/requirements.txt
APScheduler==3.10.4
bcrypt==4.1.2
Bcrypt-Flask==1.0.2
blinker==1.7.0
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.7
colorama==0.4.6
colorlog==6.7.0
contourpy==1.2.0
cycler==0.12.1
dnspython==2.4.2
docstring-parser==0.15
email-validator==2.1.0.post1
Flask==3.0.0
Flask-APScheduler==1.13.1
Flask-Bcrypt==1.0.1
Flask-Email==1.4.4
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-MySQLdb==2.0.0
Flask-Principal==0.4.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.47.0
future==0.18.3
gensim==4.3.2
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.0
greenlet==3.0.3
httplib2==0.22.0
idna==3.4
iso8601==1.1.0
isodate==0.6.1
itsdangerous==1.1.0
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
more-itertools==10.2.0
mysql-connector==2.2.9
mysql-connector-python==8.2.0
mysqlclient==2.2.1
nltk==3.8.1
numpy==1.26.3
oauthlib==3.2.2
packaging==23.2
pandas==2.1.4
pillow==10.2.0
protobuf==4.21.12
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.0.9
pyserial==3.5
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0
regex==2023.12.25
requests==2.30.0
rsa==4.9
scipy==1.11.4
seaborn==0.13.1
serial==0.0.97
six==1.16.0
smart-open==6.4.0
SQLAlchemy==2.0.25
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.0.2
websocket-client==1.6.4
Werkzeug==3.0.1
WTForms==3.1.2